# [Xamarin.Android.Build.Tasks] remove the linksrc directory