I've removed various MSBuild targets that copy assemblies into the
`obj\Debug\linksrc` directory, and instead just use
`@(ResolvedUserAssemblies)` or `@(ResolvedAssemblies)` directly

Things seem to *work*, I can build and deploy `samples\HelloWorld` and
the Xamarin.Forms app in this repo. Let's see what breaks?

The size of `tests\Xamarin.Forms-Performance-Integration\Droid\obj\`
is improved dramatically:

    Before:
    184.27 MB
    After:
    111.74 MB

~73MB savings.

These MSBuild tasks were completely removed:

      3 ms  _CopyMdbFiles                              1 calls
     43 ms  _CopyPdbFiles                              1 calls
    127 ms  _CopyIntermediateAssemblies                1 calls

On a slower machine with Windows Defender enabled, these targets are
taking:

      6 ms  _CopyMdbFiles                              1 calls
     95 ms  _CopyPdbFiles                              1 calls
    269 ms  _CopyIntermediateAssemblies                1 calls

On the slower machine, the overall time was even more drastic:

    Before:
    Time Elapsed 00:00:33.41
    After:
    Time Elapsed 00:00:29.92

With Windows Defender processes using the CPU, the overall build time
is improved by ~3.5 seconds?

## Changes for jnimarshalmethod-gen ##

Because `jnimarshalmethod-gen` edits assemblies in `linksrc` in-place,
before the linker, I added an extra `jnisrc` directory to accommodate
for this. Then when `_LinkAssembliesShrink` runs, it operates on
assemblies in the `jnisrc` directory.

We can revisit this if `$(AndroidGenerateJniMarshalMethods)` ever
becomes default.