@jonathanpeppers 2019-07-18
Unfortunately there is a problem here:
1. `jnimarshalmethod-gen.exe` operates on assemblies in-place in `linksrc`
1. We can't run this _after_ the linker, because the Linker expects it to have been run: https://github.com/xamarin/xamarin-android/blob/05344f8d43293c794f3df2b533b71d339ef031f7/src/Xamarin.Android.Build.Tasks/Linker/MonoDroid.Tuner/MonoDroidMarkStep.cs#L35

With the current implementation of `jnimarshalmethod-gen`, we need an extra copy of the .NET assemblies on disk: `linksrc`. Need to think on this...