@jonathanpeppers 2019-08-20
@radekdoulik after I've switched this around to use `-o`. I'm getting:
```
 Unhandled Exception:
  System.UnauthorizedAccessException: Access to the path "/Library/Frameworks/Mono.framework/External/xbuild-frameworks/MonoAndroid/v10.0/Mono.Android.dll" is denied.
```
This only happens because our tests on Azure DevOps are running against a system install.

Is `jnimarshalmethod-gen.exe` opening this file with write access?