@radekdoulik 2019-08-20
Indeed, we were opening the assemblies with `ReadWrite` reader parameters even when using `-o`. It should be now fixed by https://github.com/xamarin/java.interop/commit/d39699ff541005c1a95338d4a0df6c642f356e24