@jonathanpeppers 2019-08-21
@radekdoulik got past that one, thanks.

There is a new error though 😞:
```
EXEC : error : jnimarshalmethod-gen: Unable to process assembly '/Library/Frameworks/Mono.framework/External/xbuild-frameworks/MonoAndroid/v10.0/Mono.Android.dll' [/Users/vsts/agent/2.155.1/work/1/s/src/Mono.Android/Test/Mono.Android-Tests.csproj]
  Access to the path "/Library/Frameworks/Mono.framework/External/xbuild-frameworks/MonoAndroid/v10.0/Mono.Android-JniMarshalMethods.dll" is denied.
  System.UnauthorizedAccessException: Access to the path "/Library/Frameworks/Mono.framework/External/xbuild-frameworks/MonoAndroid/v10.0/Mono.Android-JniMarshalMethods.dll" is denied.
    at System.IO.FileStream..ctor (System.String path, System.IO.FileMode mode, System.IO.FileAccess access, System.IO.FileShare share, System.Int32 bufferSize, System.Boolean anonymous, System.IO.FileOptions options) [0x001b7] in <b047284b444b4c74b59771ad4346b34a>:0 
    at System.IO.FileStream..ctor (System.String path, System.IO.FileMode mode, System.IO.FileAccess access, System.IO.FileShare share, System.Int32 bufferSize, System.Boolean isAsync, System.Boolean anonymous) [0x00000] in <b047284b444b4c74b59771ad4346b34a>:0 
    at System.IO.FileStream..ctor (System.String path, System.IO.FileMode mode, System.IO.FileAccess access) [0x00000] in <b047284b444b4c74b59771ad4346b34a>:0 
    at (wrapper remoting-invoke-with-check) System.IO.FileStream..ctor(string,System.IO.FileMode,System.IO.FileAccess)
    at System.Reflection.Emit.ModuleBuilder.Save () [0x001e5] in <b047284b444b4c74b59771ad4346b34a>:0 
    at System.Reflection.Emit.AssemblyBuilder.Save (System.String assemblyFileName, System.Reflection.PortableExecutableKinds portableExecutableKind, System.Reflection.ImageFileMachine imageFileMachine) [0x0022b] in <b047284b444b4c74b59771ad4346b34a>:0 
    at System.Reflection.Emit.AssemblyBuilder.Save (System.String assemblyFileName) [0x00000] in <b047284b444b4c74b59771ad4346b34a>:0 
    at Xamarin.Android.Tools.JniMarshalMethodGenerator.App.CreateMarshalMethodAssembly (System.String path) [0x0043d] in <3859ef09ce2c46b7b24c5a9fe9afd6b4>:0 
    at Xamarin.Android.Tools.JniMarshalMethodGenerator.App.ProcessAssemblies (System.Collections.Generic.List`1[T] assemblies) [0x001b0] in <3859ef09ce2c46b7b24c5a9fe9afd6b4>:0 
```
Is this `Mono.Android-JniMarshalMethods.dll` an assembly generated that should be going into the `-o` directory? I need also make sure it makes it into the linker step and the app?