@radekdoulik 2019-08-21
> Is this Mono.Android-JniMarshalMethods.dll an assembly generated that should be going into the -o directory? I need also make sure it makes it into the linker step and the app?

Definitely, it should be generated in the directory specified by `-o`. I will fix it as well.

The assembly itself is temporary.