@radekdoulik 2019-08-21
https://github.com/xamarin/java.interop/pull/483 should fix it.