@mjo151 2019-11-21
We just upgraded to VS 2019 and this change appears to have broken obfuscation with our Xamarin.Android builds. We were performing obfuscation after the _CopyIntermediateAssemblies target, then copying our obfuscated assemblies into the linksrc directory. Due to this change, this no longer works. Can you provide suggestions on which build target we should use to perform obfuscation? Also, where does the linker look for assemblies (e.g. where should be copy our obfuscated assemblies)?