@jonathanpeppers 2019-11-21
@mjo151 the convention in MSBuild is that anything prefixed with an underscore is _private_. Meaning it might change from release to release. (MSBuild is all global variables otherwise)

https://github.com/xamarin/xamarin-android/blob/master/Documentation/guides/MSBuildBestPractices.md#naming

However, you can easily copy the files to a directory yourself and fix up the `@(ResolvedAssemblies)` and `@(ResolvedUserAssemblies)` item groups, we had to do a similar workaround for AndroidX.Migration: https://github.com/xamarin/XamarinAndroidXMigration/pull/20

You can check if `$(MonoAndroidLinkerInputDir)` is blank to determine if it is a newer Xamarin.Android version where `linksrc` is gone.